refactor(api): move 'util.rand64() -> crypto.rand64()'
